Feature/pooling backed by randomness (#34)
When connecting via a load balancer, each physical connection in a pool may be connected to different casual domains.
Thus they can each have their own set of queues and services and we need to take that into account.
We make it so that for each call we can provide the domain id so that the appserver can match on that and return a handle for which the physical connection has that exact domain id.

We record each domain id per host:port ( address) combination for all new ManagedConnections that the appserver creates.
This data can then be queried by casual-caller.
Casual-caller does this by polling, by default each 10s, and then updating the state of the world.

+++ casual/casual-network-protocol/src/main/java/se/laz/casual/network/protocol/decoding/decoders/queue/CasualDequeueRequestMessageDecoder.java
@@ -48,7 +48,7 @@ public final class CasualDequeueRequestMessageDecoder implements NetworkDecoder<
         int selectorPropertiesSize = (int) ByteUtils.readFully(channel, DequeueRequestSizes.SELECTOR_PROPERTIES_SIZE.getNetworkSize()).getLong();
         String selectorProperties = (0 == selectorPropertiesSize) ? "" : CasualMessageDecoderUtils.readString(channel, selectorPropertiesSize);
         UUID selectorId = CasualMessageDecoderUtils.readUUID(channel);
-        boolean block = (1 == (int) ByteUtils.readFully(channel, DequeueRequestSizes.BLOCK.getNetworkSize()).get());
+        boolean block = (1 == ByteUtils.readFully(channel, DequeueRequestSizes.BLOCK.getNetworkSize()).get());
         return CasualDequeueRequestMessage.createBuilder()
                                           .withExecution(execution)
                                           .withQueueName(queueName)
@@ -83,7 +83,7 @@ public final class CasualDequeueRequestMessageDecoder implements NetworkDecoder<
         currentOffset += selectorPropertiesSize;
         UUID selectorId = CasualMessageDecoderUtils.getAsUUID(Arrays.copyOfRange(bytes, currentOffset, currentOffset + DequeueRequestSizes.SELECTOR_ID_SIZE.getNetworkSize()));
         currentOffset += CommonSizes.EXECUTION.getNetworkSize();
-        boolean block = (1 == (int)ByteBuffer.wrap(bytes, currentOffset , DequeueRequestSizes.BLOCK.getNetworkSize()).get());
+        boolean block = (1 == ByteBuffer.wrap(bytes, currentOffset , DequeueRequestSizes.BLOCK.getNetworkSize()).get());
         return CasualDequeueRequestMessage.createBuilder()
                                           .withExecution(execution)
                                           .withQueueName(queueName)
